+++ src/Presentation/SmartStore.Web/Administration/Models/Catalog/CategoryTreeModel.cs
@@ -5,24 +5,15 @@ using SmartStore.Web.Framework.Mvc;
 
 namespace SmartStore.Admin.Models.Catalog
 {
-    public class CategoryListModel : ModelBase
+    public class CategoryTreeModel : ModelBase
     {
-		public CategoryListModel()
+		public CategoryTreeModel()
 		{
 			AvailableStores = new List<SelectListItem>();
 		}
 
-        [SmartResourceDisplayName("Admin.Catalog.Categories.List.SearchCategoryName")]
-        [AllowHtml]
-        public string SearchCategoryName { get; set; }
-
-		[SmartResourceDisplayName("Admin.Catalog.Categories.List.SearchAlias")]
-		public string SearchAlias { get; set; }
-
 		[SmartResourceDisplayName("Admin.Common.Store.SearchFor")]
 		public int SearchStoreId { get; set; }
 		public IList<SelectListItem> AvailableStores { get; set; }
-
-		public int GridPageSize { get; set; }
     }
 }