+++ changelog.md
@@ -7,6 +7,7 @@
 * #652 Web API: Support for file upload and multipart mime
 * New option to display product thumbnails in instant search (NOTE: the feature can be enabled via _Configuration > Settings > Catalog Settings > Search_)
 * #594 Option to prevent a message from being automatically sent
+* Category filtering by store for backend category tree and list
 
 ## SmartStore.NET 2.2.0